requirements.txt
torch==1.11.0
torchvision==0.12.0
fastapi==0.78.0
uvicorn==0.18.2
Pillow==9.2.0
numpy==1.23.1
python-multipart==0.0.5
